# remove redundant OR