@EgorBo 2023-07-17
Presumably, you can use the same trick in the main loop of these functions that currently looks like this:
```diff
-if (!AllCharsInVectorAreAscii(leftValues | rightValues))
-{
-    return false;
-}

Vector128<TRight> notEquals = ~Vector128.Equals(leftValues, rightValues);

if (notEquals != Vector128<TRight>.Zero)
{
+   if (!AllCharsInVectorAreAscii(leftValues | rightValues))
+       return false;
    // not exact match

    leftValues |= loweringMask;
    rightValues |= loweringMask;

    if (Vector128.GreaterThanAny((leftValues - vecA) & notEquals, vecZMinusA) || leftValues != rightValues)
    {
        return false; // first input isn't in [A-Za-z], and not exact match of lowered
    }
}
+else if (!AllCharsInVectorAreAscii(leftValues))
+   return false;
```

but I'm not sure if it's profitable - `or` is fast and this increases loop body's size 🙂